+++ lib/src/ui/cells/text_cell.dart
@@ -69,28 +69,24 @@ mixin TextCellState<T extends TextCell> on State<T> implements TextFieldProps {
 
   @override
   void dispose() {
-    _debounce.dispose();
-
-    _textController.dispose();
-
-    cellFocus.dispose();
-
     /**
      * Saves the changed value when moving a cell while text is being input.
      * if user do not press enter key, onEditingComplete is not called and the value is not saved.
      */
     if (_cellEditingStatus.isChanged) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        _changeValue(notify: false);
-
-        widget.stateManager.notifyListenersOnPostFrame();
-      });
+      _changeValue();
     }
 
     if (!widget.stateManager.isEditing ||
         widget.stateManager.currentColumn?.enableEditingMode != true) {
       widget.stateManager.setTextEditingController(null);
     }
+
+    _debounce.dispose();
+
+    _textController.dispose();
+
+    cellFocus.dispose();
 
     super.dispose();
   }
@@ -137,28 +133,22 @@ mixin TextCellState<T extends TextCell> on State<T> implements TextFieldProps {
     return false;
   }
 
-  void _changeValue({bool notify = true}) {
+  void _changeValue() {
     if (formattedValue == _textController.text) {
       return;
     }
 
-    widget.stateManager.changeCellValue(
-      widget.cell,
-      _textController.text,
-      notify: notify,
+    widget.stateManager.changeCellValue(widget.cell, _textController.text);
+
+    _textController.text = formattedValue;
+
+    _initialCellValue = _textController.text;
+
+    _textController.selection = TextSelection.fromPosition(
+      TextPosition(offset: _textController.text.length),
     );
 
-    if (notify) {
-      _textController.text = formattedValue;
-
-      _initialCellValue = _textController.text;
-
-      _textController.selection = TextSelection.fromPosition(
-        TextPosition(offset: _textController.text.length),
-      );
-
-      _cellEditingStatus = _CellEditingStatus.updated;
-    }
+    _cellEditingStatus = _CellEditingStatus.updated;
   }
 
   void _handleOnChanged(String value) {